corrected setters
