+++ spinn_front_end_common/interface/abstract_spinnaker_base.py
@@ -568,29 +568,6 @@ class AbstractSpinnakerBase(ConfigHandler, SimulatorInterface):
     REPORT_DISABLE_OPTS = frozenset([
         "clear_iobuf_during_run", "extract_iobuf", "extract_iobuf_during_run"])
 
-    def set_up_timings(self, machine_time_step=None, time_scale_factor=None):
-        """ Set up timings of the machine
-
-        :param machine_time_step:\
-            An explicitly specified time step for the machine.  If None,\
-            the value is read from the config
-        :param time_scale_factor:\
-            An explicitly specified time scale factor for the simulation.\
-            If None, the value is read from the config
-        """
-
-        # set up timings
-        if machine_time_step is not None:
-            self.machine_time_step = machine_time_step
-
-        if self.machine_time_step <= 0:
-            raise ConfigurationException(
-                "invalid machine_time_step {}: must greater than zero".format(
-                    self.machine_time_step))
-
-        if time_scale_factor is not None:
-            self.time_scale_factor = time_scale_factor
-
     def set_up_machine_specifics(self, hostname):
         """ Adds machine specifics for the different modes of execution
